feat: allow debug-log command against 3.x
Provides _client_ backwards capability to talk to a 3.x controller for
streaming debug-logs. This piggy backs off of the HA facade and
inspecting if there is a way to get the ControllerDetails. If there is,
it will use a connector to each individual controller, otherwise if
it isn't available, it will connect to the current controller and
stream it from mongo that way.