feat: trace config to k8s units
This fixes a bug where we weren't passing the trace config into the
k8s units agent.conf. This fixes that issue so that if any unit is
created it will have the right information. Once we have the right
config, we can be sure that during hook executions the TRACE flags
are correctly set.