chore: refactor network modelmigration import operations
During model import we have a problem, related with the order of the
import operations where the import of machines would fail because of
missing import of availability zones. This patch solves it by splitting
the network import again into LLDs that run *after* the machine import
and the subnets (plus spaces and AZs) that runs *before* machines.