feat: add tooling to view tables within transaction before rollback
We have been witnessing _a lot_ of foreign key constraint errors within
the removal domain. Also, SQLite returns really bad errors when this
happens. Simply 'foreign key constraint violation', without even telling
us what table was being modified, or what table caused the error

Also, when we see this error, the very next step is to rollback the
transaction, so the existing DumpTable helper function is inadequate, we
lose the state that resulted in this error.

As a result, these errors are very difficult to debug.

Patch our testing txnRunner for the DqliteSuite to shim the txn
function, allowing us to inject a posthook to run when the usual
function exits but before the transaction is rolled back.

This allows us to inspect the state of the database when the error was
thrown

Also, write some helper functions, particularly
DumpForeignKeysForTableSqlair, which uses a sqlair.TX to dump the
contents of all tables that have a foreign key link to the provided
table