fix: transaction rollback handling
Do not attempt to retry transaction rollbacks. The transaction is marked
as done upon the first attempt, so whether it returns an error or not
this action is not retryable.

We only log the warning if the rollback error was not due to the
transactiona already being done, which is the case for context
cancellation. In the same way, commit errors with the same cause are
ignored.

Adds the same logic used by LXD to handle a state that we can get into
(apparently) when Dqlite is performing a checkpoint operation.
We attempt a hard ROLLBACK against the DB in order not to end up with a
poisoned database.