feat: read history from multiple controllers
In HA mode, a singular controller might not have all the status history.
Instead, we need to ask each controller for their view of the history.
We need to correctly handle gracefully, if a controller node is down
and also provide compatibility if we're targetting 3.6. In 3.6 all the
data was stored in mongo, so a singular query to one controller would
be enough. In 4.0, we store the history outside of the database, because
of it's unbounded growth, storing it in the logsink files. We then
read those to see if we can locate that information. Long term, this
should be stored in loki and viewable from that perspective.