refactor(secret/state): extract shared DB query logic into helpers
This patch is the first step to remove RunAtomic from the secret domain.
It extracts the logic of each AtomicState function into a helper which takes a sqlair.TX parameter.

This will allow using them in other transactional context inside the state, while the
functionalities will be refactored and the code logic move down to the state while we remove
RunAtomic.