feat: add generator for model persistence types
For each declared version, a new package under export/types/ will be
created, and contain the generated persistence types for every table.

There is an exception to the general rule here with persistence types
being exposed outside of state. This is because we later want to
transform between export versions in the service layer.