fix: api remote caller reconnect on broken
This forces the API remote caller to reconnect when the current
connection becomes broken. The prior change caused the worker to
restart when the connection became broken. This was clean from a design
of the worker, except we lost the address we should be attempting to
connect to when we restarted. Thus, what ended up happening, is that
we would connect, become broken and then wait for an address to be
set. In other words, we would just hang.

The solution is to force a reconnect in the same worker. We can just
piggy back on sending the current connection address to the changes and
force a re-estabilished connection.