feat: subscribe to api remote caller changes
If a new connection is opened via a new connection, then allow others
to subscribe to the change. That way it's possible to keep insync with
changes to the underly controller node topology.

This will make it easy to handle changes to the control plane of
controller nodes when timeouts or termination of connections happen.