feat: controller presence worker
The controller presence worker will start up and subscribe to the
api-remote-caller worker. This will allow it to start tracking the
controller api connection plane. Any changes to this plane will also
mean that the controller presence will be updated.

If a connection becomes broken, it will mark the controller node
machine and unit as down. Once it comes back it will restart the worker
and wait for the same condition. It's sole job is to provide better
latency for notification when controller nodes are lost as sea.