feat: allow editing of pings in api monitor
For controller connections it's better for the controller plane to have
reduced ping periods. So instead of waiting a full minute, we can drop
that to something more sensible (every 30 seconds). In the worst case
you'd wait for 1 minute (ping period and timeout - 30 seconds each),
instead of 1.5 minutes. We could reduce it more, but I don't think we
need to be that active. This is the worst case. In most cases, we will
be notified a lot sooner, either by the connection gracefully closing,
or the mintor actively noticing sooner.