feat: introduce and use InsertIAASUnitState
Split application state into 2 pieces. One which covers InsertIAASUnit
functionality and the other for everything else.

The new InsertIAASUnitState will be shared with the relation domain for
creating subordinate units during EnterScope.

This approach replaces the previous use of the AddIAASSubordinateUnit
method and SubordinateCreator interface. It allows for EnterScope
including subordinate unit creation to be done in a (almost) single
transaction, removing the need to back out a relation if the subordinate
unit creation fails. RecordStatus for StatusHistory will still be done
in a separate transaction.