refactor: replace charmID struct with entityUUID
Once charmID had a simple type for the UUID, it became a duplicate of
entityUUID struct. Thus it can be replaced.

This caused some queries to have 2 different entityUUIDs, one for the
application and one for the charm, causing Prepare to fail. Created a
new type to pair these values together: applicationAndCharmUUID.