feat: attach the volume mounts for workload container
Here we consume the []FilesystemAttachmentTemplate for the workload container.

There are also adjustments how we treat a duplicated PVC. If we have found the exact same PVC then it must be from a previous workload/charm container and we want to reuse that. We still return an err if the PVC name is the same but the other contents are different.