feat(actions): add action to check for DDL patches
We do not allow DDL patch files on main. Patch files are intended to be
applied to the DDL when a controller is upgraded.

Since the main branch is used for the latest in development version,
we know there are no previous patch versions we can upgrade from. So any
DDL patches/changes should be applied directly to the DDL