refactor: start optimising provisioning info
It was observed that we're calling model info and config for every
single machine we're provisioning. This isn't the worst of it, there
is a lot more we can do, but this is a start.

Additionally, to ensure that during the whole call of the provisioning
info, we'll cache the same values from model config. This ensures that
if another operator changes model config at the same time, we will
get predicatable repeatable outcomes.

What is to follow, will be more reduction in calls, with the aim at
speed provisioning up.