feat: services and state are model scoped
We don't need nor want to pass in a model UUID into methods that don't
need them. Instead, everything is and always will be model scoped.

This fixes that by making the application state aware of the model
uuid. It's a bit more work, but is a lot cleaner in the long run.