feat: ensure we set a machine archiecture
We need to specify the machine architecture to a default architecture
if none is set. The replicates what happens in 3.6. We can also
simplify a lot of the logic based on this.

Lastly. we **DO NOT** need to decode the constraints every time we
get it, just write it to a variable.