feat: pass instance id to machine placement
If we're providing a machine placement directive, with hardware
charactistics then we must also pass the instance id along with it.
Otherwise it will think it's not provisioned. This happens on manual
machines.

Secondly, the short-circuit that was done, made a lot of the tests fail
because we didn't correctly set up the unit or applications. There isn't
a worker running to update the hardware characteristics so we have to
do it upfront, or manually place them there.