feat: start CommitHookChanges domain method
Scaffolding written for the CommitHookChanges domain method.

Implemention of save unit state charm data, as it was done in the domain
already.

CommitHookChangesArg structure to provide the data to be commited on a
hook change. The ValidateAndLeadership method validates the input and
indicates whether leadership is required for this data. Leadership is
required for Application Settings and Secret related changes.