refactor: maas storage and volume providers
This commit refactors the storage and volume providers for MAAS to fix a
bug with deploying charms in Juju 4.0. Due to a change with storage
providers and their ability to recommend storage pools and provide
default storage pools into a model it now meant that the maas storage
provider was being used to provision workload storage.

This resulted in a bug when deploying postgres or other charms requiring
storage as MAAS was being used to try and provide the storage on top of
a machine which it is not capable of doing. MAAS can only provide
storage to support root disks of new machines being created in a model.

This commit stops the storage provider interface of the MAAS environ
from recommending the MAAS storage pool for use. This will flow on and
stop workloads from using the storage pool as a default.

On top of this I have also made the storage provider also return false
for supporting both volumes and filesystems. This will make it so the
storage provider can never be used for charm storage.