fix(ci): metrics tests didn't check the right logger
This patch refactors the `metrics.sh` test script to check the correct logger.
Juju 4 has refactored several part of the charm refresh workload, moving code to a worker which
use directly the domain, without passing through a facade.
So the logs weren't in the controller log anymore, but instead directly in model logs, under another
logger (not the facade logger).

As a fly-by, there is a minor refactoring of the way we select the base of the ubuntu image to
define it once in run_model_metrics, which helps maintainability when we want to upgrade the base
(subordinate relation requires that all application in the test use the same base).

In short:
- Replacing hardcoded model names with the `$testname` variable.
- Updating logging configurations to align with the `$testname` model context.
- Renaming `charmrevisionupdater` to `charmrevisioner` for consistency with worker naming.
- Using a variable for the Ubuntu base to streamline version changes.