fix: a parsing issue with tags on maas storage provider
The MAAS storage provider had parsing logic for validating tags supplied
to the storage provider. The validation logic was only running on tags
supplied as a string and not on slices of string tags which are also
supported by the provider.

The provider was also not asserting the correct type of the coerced
value failing to parse slices of tags at all.

This commit applies the same validation logic to all forms of tags and
introduces a new test suite to assert the storage provider in it's own
space and file.