refactor: serve connection
This refactors the serveConn method to ensure that any prerequisites
for a connection are satisified before attempting the facade
connections. One of the main changes, is that the apiHandler no longer
checks for the model, that instead is done in the serveConn method.
This reduces potential starvation attacks against the controller, that
want access an invalid model. Previously, we would try and get the
domain services, tracer, objectstore, and more, which aren't exactly
cheap to access.

It should be noted that if we can't access any of the prerequisite
dependencies previously, that would only log an error. This could leave
the connection open and tax the controller. Instead, the change here
will explicitly close the connection and free critical resources.

This a fundamental change and we should ensure correctness.