refactor: use Name and UUID types in CommitHookChangesArg
We cannot truely validate if the values are not in a type. An internal
package version of CommitHookChangesArg uses scalare types between the
service and state pacakges of the domain.

The unit's UUID will be found in state, rather than provided by the
facade.

Renames the struct's Validate method to ValidateAndHasChanges to fully
encompass its functionality.